+++ tests/unit/test_config.py
@@ -1,4 +1,4 @@
-import os.path
+import logging
 import oraclebmc
 
 import pytest
@@ -25,15 +25,13 @@ def test_load_default_profile():
     assert config["user"] == HARDCODED_USER
     assert config["fingerprint"] == HARDCODED_FINGERPRINT
     assert config["region"] == HARDCODED_REGION
-    assert config["key_file"] == os.path.expanduser(HARDCODED_KEYFILE)
+    # key file is expanded by the signer, not config
+    assert config["key_file"] == HARDCODED_KEYFILE
 
 
 def test_child_profile():
     config = oraclebmc.config.from_file(
         file_location=get_resource_path('config'), profile_name='DEBUG')
-
-    # check some default properties
-    assert config["verify_ssl"]
 
     # check properties inherited from the default profile
     assert config["tenancy"] == HARDCODED_TENANCY
@@ -44,7 +42,6 @@ def test_child_profile():
     # check properties overridden by the specified profile
     assert config["log_requests"]
     assert config["additional_user_agent"]
-    assert config["endpoint"] == HARDCODED_ENDPOINT
 
 
 def test_extra_config_values():
@@ -67,9 +64,36 @@ def test_profile_not_found():
     assert 'does_not_exist' in str(excinfo.value)
 
 
-def test_invalid_region():
-    with pytest.raises(ValueError) as excinfo:
-        oraclebmc.config.from_file(
-            get_resource_path('config'),
-            profile_name='INVALID_REGION')
-    assert "Unknown region 'whoami'" in str(excinfo.value)
+def test_unknown_region(caplog):
+    config = oraclebmc.config.from_file(
+        get_resource_path('config'),
+        profile_name='UNKNOWN_REGION')
+
+    # Creating a config object doesn't log anything
+    assert not caplog.records
+
+    # Creating the client (which builds an endpoint) fires a warning
+    oraclebmc.clients.IdentityClient(config)
+    assert caplog.record_tuples == [
+        ("oraclebmc.regions", logging.WARN,
+         "Using unknown region '{}' to build service endpoint for 'identity'".format(HARDCODED_ENDPOINT))
+    ]
+
+
+def test_missing_required():
+    config = {
+        "user": "malformed-user",
+        "tenancy": "malformed-tenancy",
+        # missing "region"
+        # valid fingerprint
+        "fingerprint": "00:00:00:00:00:00:00:00:00:00:00:00:00:00:00:00",
+        "key_file": "~/.oraclebmc/config"
+    }
+    with pytest.raises(oraclebmc.exceptions.InvalidConfig) as excinfo:
+        oraclebmc.config.validate_config(config)
+
+    assert excinfo.value.errors == {
+        "user": "malformed",
+        "tenancy": "malformed",
+        "region": "missing"
+    }